requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
rapidfuzz>=3.9.0